moved and renamed test files for clarity #68
