+++ lib/test-helpers/failing-module.js
