+++ lib/test-helpers/tester.js
