+++ test/ejector/integration-gearmand.js
@@ -5,7 +5,7 @@ var gearman = require('gearman-coffee')
   , sinon = require('sinon')
   , sinonChai = require('sinon-chai')
   , async = require('async')
-  , spawn = require('../lib/spawn');
+  , spawn = require('../../lib/test-helpers/spawn');
 
 chai.should();
 chai.use(sinonChai);
