+++ test/extras-for-adapters/unit-composite.js
